Remove hardcoded eu region API url
The correct url to call can be fetched from the sdkToken used to initialize
the OnfidoAuth sdk.

This commit removes the EU BaseURL that was being used allowing the auth
sdk to be used in any given region.

https://onfido.atlassian.net/browse/BIOMANU-1341
